fix(nossas): rename base classes to render UI plugins
